requirements.txt
streamlit
google-generativeai
langchain
langchain-community
faiss-cpu
pypdf
pymupdf
python-dotenv
